src/serving/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.5.0
xgboost==2.1.0
joblib==1.4.0
loguru==0.7.2
google-cloud-storage==2.18.0
httpx==0.27.0
